+++ test/Telegram.Bot.Tests.Unit/User/UserTests.cs
@@ -1,14 +1,15 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using Newtonsoft.Json;
 using Telegram.Bot.Types;
 using Xunit;
 
-namespace Telegram.Bot.Tests.Unit
+namespace Telegram.Bot.Tests.Unit.UserTests
 {
+    [Collection("User Tests")]
     public class UserTests
     {
-        [Fact]
+        [Fact(DisplayName ="Should Convert User Object to String")]
         public void Should_ToString()
         {
             Assert.Equal("@alicebot (12345)", new User
@@ -35,7 +36,7 @@ namespace Telegram.Bot.Tests.Unit
             }.ToString());
         }
 
-        [Fact]
+        [Fact(DisplayName = "Should Deserialize User")]
         public void Should_Add_To_Dict()
         {
             const int id1 = 12345;